+++ src/main/java/com/kustacks/kuring/worker/update/notice/DepartmentNoticeUpdater.java
@@ -1,11 +1,7 @@
 package com.kustacks.kuring.worker.update.notice;
 
 import com.kustacks.kuring.category.domain.Category;
-import com.kustacks.kuring.common.dto.NoticeMessageDto;
-import com.kustacks.kuring.common.error.ErrorCode;
-import com.kustacks.kuring.common.error.InternalLogicException;
 import com.kustacks.kuring.common.firebase.FirebaseService;
-import com.kustacks.kuring.common.firebase.exception.FirebaseMessageSendException;
 import com.kustacks.kuring.notice.domain.DepartmentName;
 import com.kustacks.kuring.notice.domain.DepartmentNotice;
 import com.kustacks.kuring.notice.domain.DepartmentNoticeRepository;
@@ -44,7 +40,8 @@ public class DepartmentNoticeUpdater {
 
     private static long startTime = 0L;
 
-    @Scheduled(cron = "0 10/20 8-18 * * *", zone = "Asia/Seoul") // 학교 공지는 오전 8:10 ~ 오후 6:55분 사이에 20분마다 업데이트 된다.
+    //@Scheduled(fixedRate = 30, timeUnit = TimeUnit.DAYS)
+    //@Scheduled(cron = "0 10/20 8-18 * * *", zone = "Asia/Seoul") // 학교 공지는 오전 8:10 ~ 오후 6:55분 사이에 20분마다 업데이트 된다.
     public void update() {
         log.info("******** 학과별 최신 공지 업데이트 시작 ********");
         startTime = System.currentTimeMillis();
@@ -53,7 +50,7 @@ public class DepartmentNoticeUpdater {
             CompletableFuture
                     .supplyAsync(() -> updateDepartmentAsync(deptInfo, DeptInfo::scrapLatestPageHtml), noticeUpdaterThreadTaskExecutor)
                     .thenApply(scrapResults -> compareLatestAndUpdateDB(scrapResults, deptInfo.getDeptName()))
-                    .thenAccept(this::sendNotificationByFcm);
+                    .thenAccept(firebaseService::sendNotificationByFcm);
         }
     }
 
@@ -141,7 +138,9 @@ public class DepartmentNoticeUpdater {
 
         departmentNoticeRepository.saveAllAndFlush(newNotices);
 
-        departmentNoticeRepository.deleteAllByIdsAndDepartment(departmentNameEnum, deletedNoticesArticleIds);
+        if(!deletedNoticesArticleIds.isEmpty()) {
+            departmentNoticeRepository.deleteAllByIdsAndDepartment(departmentNameEnum, deletedNoticesArticleIds);
+        }
     }
 
     private List<DepartmentNotice> filteringSoonSaveNotice(List<CommonNoticeFormatDto> scrapResults, List<Integer> savedArticleIds, DepartmentName departmentNameEnum, boolean important) {
@@ -188,26 +187,5 @@ public class DepartmentNoticeUpdater {
                 .category(category)
                 .departmentName(departmentNameEnum)
                 .build();
-    }
-
-    private void sendNotificationByFcm(List<DepartmentNotice> departmentNoticeList) {
-        List<NoticeMessageDto> departmentDtoLists = departmentNoticeList.stream()
-                .map(NoticeMessageDto::from)
-                .collect(Collectors.toList());
-
-        try {
-            firebaseService.sendNoticeMessageList(departmentDtoLists);
-            log.info("FCM에 정상적으로 메세지를 전송했습니다.");
-            log.info("전송된 공지 목록은 다음과 같습니다.");
-            for (DepartmentNotice notice : departmentNoticeList) {
-                log.info("아이디 = {}, 날짜 = {}, 카테고리 = {}, 제목 = {}", notice.getArticleId(), notice.getPostedDate(), notice.getDepartmentName(), notice.getSubject());
-            }
-        } catch (FirebaseMessageSendException e) {
-            log.error("새로운 공지의 FCM 전송에 실패했습니다.");
-            throw new InternalLogicException(ErrorCode.FB_FAIL_SEND, e);
-        } catch (Exception e) {
-            log.error("새로운 공지를 FCM에 보내는 중 알 수 없는 오류가 발생했습니다.");
-            throw new InternalLogicException(ErrorCode.UNKNOWN_ERROR, e);
-        }
     }
 }
